Altera 'long' para 'int32_t' (source)

+++ source/activex/c_activex.c
@@ -405,7 +405,7 @@ static ULONG STDMETHODCALLTYPE Invoke(IEventHandler *this, DISPID dispid, REFIID
             *((&(params->rgvarg[iArg - i]))->n1.n2.n3.piVal) = (int16_t)hb_itemGetNI(pItemArray[i - 1]);
             break;
           case VT_I4 | VT_BYREF:
-            *((&(params->rgvarg[iArg - i]))->n1.n2.n3.plVal) = (long)hb_itemGetNL(pItemArray[i - 1]);
+            *((&(params->rgvarg[iArg - i]))->n1.n2.n3.plVal) = (int32_t)hb_itemGetNL(pItemArray[i - 1]);
             break;
           case VT_R4 | VT_BYREF:
             *((&(params->rgvarg[iArg - i]))->n1.n2.n3.pfltVal) = (float)hb_itemGetND(pItemArray[i - 1]);
